+++ online-banking-sk/src/main/java/com/adyen/checkout/onlinebankingsk/OnlineBankingSKComponent.kt
@@ -18,6 +18,7 @@ import com.adyen.checkout.core.log.LogUtil
 import com.adyen.checkout.core.log.Logger
 import com.adyen.checkout.onlinebankingcore.OnlineBankingComponent
 import com.adyen.checkout.onlinebankingcore.internal.ui.OnlineBankingDelegate
+import com.adyen.checkout.onlinebankingsk.internal.provider.OnlineBankingSKComponentProvider
 
 class OnlineBankingSKComponent internal constructor(
     delegate: OnlineBankingDelegate<OnlineBankingSKPaymentMethod>,
